fix: rename package org.sunbird.mangers -> org.sunbird.managers
The directory and package declaration contained a typo ('mangers' instead
of 'managers'). This caused the file to sit in the wrong package path,
making it inconsistent with the surrounding codebase.

Changes:
- Moved FrameworkManager.scala from sunbird/mangers/ to sunbird/managers/
- Updated package declaration in FrameworkManager.scala
- Updated import in FrameworkActor.scala
- Updated import in CategoryActor.scala
- Updated import in FrameworkManagerTest.scala
